+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.8.2]
+
+### Changed
+- The `kill` command has been remodeled. Instead of always sending an `SigKill` it now allows you to specify the signal that should be sent (default is `SIGTERM`).  
+    By default the signal will be sent to the processes spawned by the `shell` parent process. The new `-a/--all` flag is provided in case you want to send the signal to the parent process as well.  
+    Available signals can be viewed with `pueue kill -h` under the `-s` flag. Either the int `15` the full name `sigterm/SIGTERM` or the abbreviation `term/TERM` can be used.  
+- The `stop` command has been removed, as it is now covered by the `kill` command
+
 
 ## [0.8.1]